requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python-headless
vtracer
rembg
onnxruntime